vyatta-cfg-system: remove quotes from SysContact and SysLocation
When the values for SysContact, SysLocation and SysDecr are written to
the snmpd.conf file, they are enclosed in backslash escaped quotes.
This isn't part of the syntax for snmpd.conf and looks wrong / could
break snmp related things, the example snmpd.conf files show these
values as unquoted.

This patch corrects the output and adds validation for the supplied
values, limiting them to a maximum of 255 printable characters
(RFC 3418).

The documentation also needs amending to take this into account.

Bug #261 http://bugzilla.vyos.net/show_bug.cgi?id=261