vyatta-cfg-system: fix for deleting encrypted password breaks config
You can't create a user without a password, but once this user is
created, you can then delete the encrypted password.  However, deleting
the encrypted password from the config doesn't actually remove the
password on the system and in addition breaks the config on reboot,
resulting in the system login node being entirely absent.

This fix ensures that the encrypted password is set to "!" when the
config node is deleted, removing the password login capability for this
account but still allowing SSH RSA key login.

Bug #336 http://bugzilla.vyos.net/show_bug.cgi?id=336